+++ Mods ChoGGi/Mark Deposit Ground/metadata.lua
@@ -1,22 +1,25 @@
 return PlaceObj("ModDef", {
 	"title", "Mark Deposit Ground",
 	"id", "ChoGGi_MarkDepositGround",
-	"steam_id", "1555446081",
 	"pops_any_uuid", "3b37cc32-66d0-4233-8366-9038691ebe0e",
+	"steam_id", "1555446081",
 	"lua_revision", 1007000, -- Picard
-	"version", 8,
+	"version", 9,
 	"version_major", 0,
-	"version_minor", 8,
+	"version_minor", 9,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {
 		"Code/Script.lua",
 	},
 	"has_options", true,
-	"description", [[Marks the ground around deposits so you can turn off the ugly signs and still see where they are.
+	"description", [[
+Marks the ground around deposits so you can turn off the ugly signs and still see where they are.
 Marks are sized depending on max amount.
 
-Includes a Mod Option to always hide the signs (you can still select the invisible sign and see the amount).
-An option to change all anomalies to the greenman model (suggestions for a different model?).
-And one more to show signs during construction mode.]],
+Mod Options:
+Hide Signs: Hide signs on the map (pressing "I" will not toggle them).
+Alien Signs: Change anomaly signs to aliens.
+Construction Signs: Signs are visible during construction.
+]],
 })